Debt - extract Continue On picker and contrib into separate class (#163002)
Debt - extract picker and contrib into separate class
